Module voicevox_core_rs::api

Structs

CPointerWrap

Holds the pointer generated by the Voicevox core and performs memory management through RAII.

CStrWrap

Holds the string generated by the Voicevox core and performs memory management through RAII.

VoicevoxAudioQueryOptions

Audio query のオプション

VoicevoxCore

Provides functionality of Voicevox Core.

VoicevoxInitializeOptions

初期化オプション

VoicevoxSynthesisOptions

voicevox_synthesis のオプション

VoicevoxTtsOptions

テキスト音声合成オプション

__fsid_t

Enums

AccelerationMode

Hardware acceleration mode

ResultCode

Enum that represents the result of a Voicevox operation.

Constants

INT8_MAX

INT8_MIN

INT16_MAX

INT16_MIN

INT32_MAX

INT32_MIN

INTPTR_MAX

INTPTR_MIN

INT_FAST8_MAX

INT_FAST8_MIN

INT_FAST16_MAX

INT_FAST16_MIN

INT_FAST32_MAX

INT_FAST32_MIN

INT_LEAST8_MAX

INT_LEAST8_MIN

INT_LEAST16_MAX

INT_LEAST16_MIN

INT_LEAST32_MAX

INT_LEAST32_MIN

PTRDIFF_MAX

PTRDIFF_MIN

SIG_ATOMIC_MAX

SIG_ATOMIC_MIN

SIZE_MAX

UINT8_MAX

UINT16_MAX

UINT32_MAX

UINTPTR_MAX

UINT_FAST8_MAX

UINT_FAST16_MAX

UINT_FAST32_MAX

UINT_LEAST8_MAX

UINT_LEAST16_MAX

UINT_LEAST32_MAX

VoicevoxAccelerationMode_VOICEVOX_ACCELERATION_MODE_AUTO

実行環境に合った適切なハードウェアアクセラレーションモードを選択する

VoicevoxAccelerationMode_VOICEVOX_ACCELERATION_MODE_CPU

ハードウェアアクセラレーションモードを“CPU“に設定する

VoicevoxAccelerationMode_VOICEVOX_ACCELERATION_MODE_GPU

ハードウェアアクセラレーションモードを“GPU“に設定する

VoicevoxResultCode_VOICEVOX_RESULT_EXTRACT_FULL_CONTEXT_LABEL_ERROR

コンテキストラベル出力に失敗した

VoicevoxResultCode_VOICEVOX_RESULT_GET_SUPPORTED_DEVICES_ERROR

サポートされているデバイス情報取得に失敗した

VoicevoxResultCode_VOICEVOX_RESULT_GPU_SUPPORT_ERROR

GPUモードがサポートされていない

VoicevoxResultCode_VOICEVOX_RESULT_INFERENCE_ERROR

推論に失敗した

VoicevoxResultCode_VOICEVOX_RESULT_INVALID_AUDIO_QUERY_ERROR

無効なAudioQuery

VoicevoxResultCode_VOICEVOX_RESULT_INVALID_MODEL_INDEX_ERROR

無効なmodel_indexが指定された

VoicevoxResultCode_VOICEVOX_RESULT_INVALID_SPEAKER_ID_ERROR

無効なspeaker_idが指定された

VoicevoxResultCode_VOICEVOX_RESULT_INVALID_UTF8_INPUT_ERROR

無効なutf8文字列が入力された

VoicevoxResultCode_VOICEVOX_RESULT_LOAD_METAS_ERROR

メタ情報読み込みに失敗した

VoicevoxResultCode_VOICEVOX_RESULT_LOAD_MODEL_ERROR

modelの読み込みに失敗した

VoicevoxResultCode_VOICEVOX_RESULT_NOT_LOADED_OPENJTALK_DICT_ERROR

open_jtalk辞書ファイルが読み込まれていない

VoicevoxResultCode_VOICEVOX_RESULT_OK

成功

VoicevoxResultCode_VOICEVOX_RESULT_PARSE_KANA_ERROR

aquestalk形式のテキストの解析に失敗した

VoicevoxResultCode_VOICEVOX_RESULT_UNINITIALIZED_STATUS_ERROR

ステータスが初期化されていない

WINT_MAX

WINT_MIN

_ATFILE_SOURCE

_BITS_STDINT_INTN_H

_BITS_STDINT_UINTN_H

_BITS_TIME64_H

_BITS_TYPESIZES_H

_BITS_TYPES_H

_BITS_WCHAR_H

_DEFAULT_SOURCE

_FEATURES_H

_POSIX_C_SOURCE

_POSIX_SOURCE

_STDC_PREDEF_H

_STDINT_H

_SYS_CDEFS_H

__FD_SETSIZE

__GLIBC_MINOR__

__GLIBC_USE_DEPRECATED_GETS

__GLIBC_USE_DEPRECATED_SCANF

__GLIBC_USE_IEC_60559_BFP_EXT

__GLIBC_USE_IEC_60559_BFP_EXT_C2X

__GLIBC_USE_IEC_60559_FUNCS_EXT

__GLIBC_USE_IEC_60559_FUNCS_EXT_C2X

__GLIBC_USE_IEC_60559_TYPES_EXT

__GLIBC_USE_ISOC2X

__GLIBC_USE_LIB_EXT2

__GLIBC__

__GNU_LIBRARY__

__HAVE_GENERIC_SELECTION

__INO_T_MATCHES_INO64_T

__LONG_DOUBLE_USES_FLOAT128

__OFF_T_MATCHES_OFF64_T

__RLIM_T_MATCHES_RLIM64_T

__STATFS_MATCHES_STATFS64

__STDC_IEC_559_COMPLEX__

__STDC_IEC_559__

__STDC_ISO_10646__

__SYSCALL_WORDSIZE

__TIMESIZE

__USE_ATFILE

__USE_FORTIFY_LEVEL

__USE_ISOC11

__USE_ISOC95

__USE_ISOC99

__USE_MISC

__USE_POSIX

__USE_POSIX2

__USE_POSIX199309

__USE_POSIX199506

__USE_POSIX_IMPLICITLY

__USE_XOPEN2K

__USE_XOPEN2K8

__WORDSIZE

__WORDSIZE_TIME64_COMPAT32

__bool_true_false_are_defined

__glibc_c99_flexarr_available

false_

true_

Functions

voicevox_audio_query⚠

voicevox_audio_query_json_free⚠

voicevox_decode⚠

voicevox_decode_data_free⚠

voicevox_error_result_to_message⚠

voicevox_finalize⚠

voicevox_get_metas_json⚠

voicevox_get_supported_devices_json⚠

voicevox_get_version⚠

voicevox_initialize⚠

voicevox_is_gpu_mode⚠

voicevox_is_model_loaded⚠

voicevox_load_model⚠

voicevox_make_default_audio_query_options⚠

voicevox_make_default_initialize_options⚠

voicevox_make_default_synthesis_options⚠

voicevox_make_default_tts_options⚠

voicevox_predict_duration⚠

voicevox_predict_duration_data_free⚠

voicevox_predict_intonation⚠

voicevox_predict_intonation_data_free⚠

voicevox_synthesis⚠

voicevox_tts⚠

voicevox_wav_free⚠

Type Definitions

AudioQueryOptions

InitializeOptions

SynthesisOptions

TtsOptions

VoicevoxAccelerationMode

VoicevoxResultCode

__blkcnt64_t

__blkcnt_t

__blksize_t

__caddr_t

__clock_t

__clockid_t

__daddr_t

__dev_t

__fsblkcnt64_t

__fsblkcnt_t

__fsfilcnt64_t

__fsfilcnt_t

__fsword_t

__gid_t

__id_t

__ino64_t

__ino_t

__int8_t

__int16_t

__int32_t

__int64_t

__int_least8_t

__int_least16_t

__int_least32_t

__int_least64_t

__intmax_t

__intptr_t

__key_t

__loff_t

__mode_t

__nlink_t

__off64_t

__off_t

__pid_t

__quad_t

__rlim64_t

__rlim_t

__sig_atomic_t

__socklen_t

__ssize_t

__suseconds_t

__syscall_slong_t

__syscall_ulong_t

__time_t

__timer_t

__u_char

__u_int

__u_long

__u_quad_t

__u_short

__uid_t

__uint8_t

__uint16_t

__uint32_t

__uint64_t

__uint_least8_t

__uint_least16_t

__uint_least32_t

__uint_least64_t

__uintmax_t

__useconds_t

int_fast8_t

int_fast16_t

int_fast32_t

int_fast64_t

int_least8_t

int_least16_t

int_least32_t

int_least64_t

intmax_t

uint_fast8_t

uint_fast16_t

uint_fast32_t

uint_fast64_t

uint_least8_t

uint_least16_t

uint_least32_t

uint_least64_t

uintmax_t