Struct voicevox_core_rs::api::VoicevoxCore

pub struct VoicevoxCore;

Provides functionality of Voicevox Core.

Implementations

impl VoicevoxCore

pub fn new(opt: InitializeOptions) -> Result<Self, ResultCode>

pub fn new_from_options(
acceleration_mode: AccelerationMode,
cpu_num_threads: u16,
load_all_models: bool,
open_jtalk_dict_dir: &CStr
) -> Result<Self, ResultCode>

impl VoicevoxCore

pub fn make_default_initialize_options() -> InitializeOptions

pub fn make_default_tts_options() -> TtsOptions

pub fn get_version() -> &'static str

pub fn get_metas_json() -> &'static str

Returns the metadata of all the available voice models in JSON format.

Returns

Returns a string representing the metadata of all available voice models in JSON format.

pub fn get_supported_devices_json() -> &'static str

Returns the list of devices supported by Voicevox in JSON format.

Returns

Returns a string representing the list of devices supported by Voicevox in JSON format.

pub fn load_model(&self, speaker_id: u32) -> Result<(), ResultCode>

Loads a model for the specified speaker ID.

Arguments

speaker_id - The ID of the speaker to load the model for.

Returns

If the model was loaded successfully, returns Ok(()). If an error occurred, returns a ResultCode enum value representing the error.

pub fn is_gpu_mode(&self) -> bool

Returns a boolean value indicating whether the current process is running in GPU mode.

Returns

Returns true if the current process is running in GPU mode, and false otherwise.

pub fn is_model_loaded(&self, speaker_id: u32) -> bool

Returns a boolean value indicating whether a voice model with the specified speaker ID has been loaded.

Arguments

speaker_id - The ID of the speaker to check for.

Returns

Returns true if a voice model with the specified speaker ID has been loaded, and false otherwise.

pub fn tts_simple(&self, text: &str, speaker_id: u32) -> Result<Wav, ResultCode>

This function generates a WAV file with the result of text-to-speech synthesis using Voicevox Core. This is simple version of VoicevoxCore::tts.

Arguments

text - The text to be synthesized.

speaker_id - The ID of the speaker to be used for the synthesis.

Returns

Wav if the synthesis succeeds.

An error code otherwise.

pub fn tts(
&self,
text: &str,
speaker_id: u32,
options: TtsOptions
) -> Result<Wav, ResultCode>

This function generates a WAV file with the result of text-to-speech synthesis using Voicevox Core.

Arguments

text - The text to be synthesized.

speaker_id - The ID of the speaker to be used for the synthesis.

options - The options for the synthesis.

Returns

Wav if the synthesis succeeds.

An error code otherwise.

pub fn error_result_to_message(result_code: ResultCode) -> &'static str

Converts a ResultCode to an error message.

Arguments

result_code - A ResultCode to convert.