Struct voicevox_core_rs::api::VoicevoxCore

pub struct VoicevoxCore;

Provides functionality of Voicevox Core.

Implementations

impl VoicevoxCore

pub fn new(opt: InitializeOptions) -> Result<Self, ResultCode>

pub fn new_from_options(
acceleration_mode: AccelerationMode,
cpu_num_threads: u16,
load_all_models: bool,
open_jtalk_dict_dir: &CStr
) -> Result<Self, ResultCode>

impl VoicevoxCore

pub fn make_default_initialize_options() -> InitializeOptions

pub fn make_default_tts_options() -> TtsOptions

pub fn make_default_audio_query_options() -> AudioQueryOptions

pub fn make_default_synthesis_options() -> SynthesisOptions

pub fn get_version() -> &'static str

pub fn get_metas_json() -> &'static str

Returns the metadata of all the available voice models in JSON format.

Returns

Returns a string representing the metadata of all available voice models in JSON format.

pub fn get_supported_devices_json() -> &'static str

Returns the list of devices supported by Voicevox in JSON format.

Returns

Returns a string representing the list of devices supported by Voicevox in JSON format.

pub fn load_model(&self, speaker_id: u32) -> Result<(), ResultCode>

Loads a model for the specified speaker ID.

Arguments

speaker_id - The ID of the speaker to load the model for.

Returns

If the model was loaded successfully, returns Ok(()). If an error occurred, returns a ResultCode enum value representing the error.

pub fn is_gpu_mode(&self) -> bool

Returns a boolean value indicating whether the current process is running in GPU mode.

Returns

Returns true if the current process is running in GPU mode, and false otherwise.

pub fn is_model_loaded(&self, speaker_id: u32) -> bool

Returns a boolean value indicating whether a voice model with the specified speaker ID has been loaded.

Arguments

speaker_id - The ID of the speaker to check for.

Returns

Returns true if a voice model with the specified speaker ID has been loaded, and false otherwise.

pub fn predict_duration(
&self,
phoneme_vector: &[i64],
speaker_id: u32
) -> Result<CPointerWrap<f32>, ResultCode>

Predicts the duration of each frame of a phoneme vector.

Arguments

phoneme_vector - A slice of the phoneme vector.

speaker_id - The ID of the speaker to synthesize speech for.

Returns

CPointerWrap if the synthesis succeeds.

An error code otherwise.

pub fn predict_intonation(
&self,
vowel_phoneme_vector: &[i64],
consonant_phoneme_vector: &[i64],
start_accent_vector: &[i64],
end_accent_vector: &[i64],
start_accent_phrase_vector: &[i64],
end_accent_phrase_vector: &[i64],
speaker_id: u32
) -> Result<CPointerWrap<f32>, ResultCode>

Predict intonation of a voice segment from given phoneme vectors.

This function predicts intonation for a voice segment from given phoneme vectors. The predicted data is returned as a dynamically allocated array of f32.

Arguments

vowel_phoneme_vector - A slice of i64 representing vowel phoneme vectors.

consonant_phoneme_vector - A slice of i64 representing consonant phoneme vectors.

start_accent_vector - A slice of i64 representing start accent vectors.

end_accent_vector - A slice of i64 representing end accent vectors.

start_accent_phrase_vector - A slice of i64 representing start accent phrase vectors.

end_accent_phrase_vector - A slice of i64 representing end accent phrase vectors.

speaker_id - An u32 representing speaker ID.

Returns

On success, returns Ok with a CPointerWrap<f32> containing the predicted intonation data.

On failure, returns Err with a ResultCode indicating the reason of failure.

pub fn decode(
&self,
phoneme_vectors: &[f32],
f0: &[f32],
speaker_id: u32
) -> Result<CPointerWrap<f32>, ResultCode>

Decode from phoneme and F0 vectors

Arguments

phoneme_vectors: Phoneme vector

f0: F0 vector

speaker_id: Speaker ID

Returns

If successful, returns a Result wrapped around a CPointerWrap object that holds the synthesized audio waveform data.

If unsuccessful, returns a Result wrapped around a ResultCode.

pub fn voicevox_synthesize(
audio_query: &str,
speaker_id: u32,
options: SynthesisOptions
) -> Result<CPointerWrap<u8>, ResultCode>

Synthesizes speech from a Voicevox audio query.

Arguments

audio_query: A &str containing the Voicevox audio query in JSON format.

speaker_id: An unsigned 32-bit integer representing the speaker ID to be used for the synthesis.

options: A struct of type VoicevoxSynthesisOptions representing the synthesis options.

Returns

CPointerWrap<u8> containing the synthesized audio in WAV format if synthesis succeeds.

An error code otherwise.

pub fn audio_query(
text: &str,
speaker_id: u32,
options: AudioQueryOptions
) -> Result<CStrWrap, ResultCode>

Sends a text query to Voicevox API to retrieve audio query information.

Arguments

text - A string slice containing the input text to be synthesized.

speaker_id - An unsigned 32-bit integer representing the speaker ID to be used for the synthesis.

options - A struct of type VoicevoxAudioQueryOptions representing the synthesis options.

Returns

audio query formatted in json format if the synthesis succeeds.

An error code otherwise.

pub fn tts_simple(
&self,
text: &str,
speaker_id: u32
) -> Result<CPointerWrap<u8>, ResultCode>

This function generates a WAV file with the result of text-to-speech synthesis using Voicevox Core. This is simple version of VoicevoxCore::tts.

Arguments

text - The text to be synthesized.

speaker_id - The ID of the speaker to be used for the synthesis.

Returns

CPointerWrap if the synthesis succeeds.

An error code otherwise.

pub fn tts(
&self,
text: &str,
speaker_id: u32,
options: TtsOptions
) -> Result<CPointerWrap<u8>, ResultCode>

This function generates a WAV file with the result of text-to-speech synthesis using Voicevox Core.

Arguments

text - The text to be synthesized.

speaker_id - The ID of the speaker to be used for the synthesis.

options - The options for the synthesis.

Returns

CPointerWrap if the synthesis succeeds.

An error code otherwise.

pub fn error_result_to_message(result_code: ResultCode) -> &'static str

Converts a ResultCode to an error message.

Arguments

result_code - A ResultCode to convert.